Update scrabbleGame.py
At the moment I've managed to complete the task whereby it will find the top three words in the dictionary. 
The user can also choose to fill a small list with five words which the system will score. 

Future update: Creating a 'bag' of random letters which the system will try to make into words and play itself at Scrabble.