Solved Challenge 11:
See generator.py
Used pathlib for easy cross-platform development and usage of
glob syntax.
Used  syntax as exercise.
Instead of manually counting the elements, I used Counter
class from collections as this is a perfect example on when
it can be used. It works on the iterator similar to a manual
loop and automatically orders the output from highest
to lowest count if the method most_common() is used.